生成式人工智能与技术应用/LCN/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2
langchain>=0.1.0
python-dotenv>=0.19.0
transformers>=4.30.0
datasets>=2.12.0
matplotlib>=3.4.3
seaborn>=0.11.2 
langchain_community>=0.0.20